Update program.py
This is the latest state of the FB Marketing codes. It runs successfully and can manually create a Custom Audience and upload users (10k at a time) to the specified Custom Audience.

To-Do: Need to refactor the logic to add http exception handling when it makes an api 'post' call to FB (audience.add_users)

Challenges:  Not able to get the status id, or http status code from the session. Not sure how to do it at the moment....